+++ src/common/layouts/AuthLayout/hooks/useHeaderDescription.ts
