+++ src/common/layouts/AuthLayout/utils/getHeaderDescription.ts
